Port bundle and string to sys crate
