+++ core-foundation-sys/src/lib.rs
@@ -1,7 +1,9 @@
-#![allow(non_snake_case, non_camel_case_types)]
+#![allow(non_snake_case, non_camel_case_types, non_upper_case_globals)]
 
 extern crate libc;
 
 pub mod array;
 pub mod base;
+pub mod bundle;
 pub mod number;
+pub mod string;
