+++ core-foundation/src/runloop.rs
@@ -14,11 +14,12 @@ use core_foundation_sys::base::{CFAllocatorRef, CFIndex, CFRelease};
 use core_foundation_sys::base::{CFTypeID, CFHashCode, mach_port_t};
 use core_foundation_sys::base::{kCFAllocatorDefault, Boolean, CFOptionFlags};
 use core_foundation_sys::array::{CFArrayRef};
+use core_foundation_sys::string::CFStringRef;
 use libc::c_void;
 use std::mem;
 
 use base::{TCFType};
-use string::{CFString, CFStringRef};
+use string::{CFString};
 use date::{CFAbsoluteTime, CFTimeInterval};
 
 pub struct CFRunLoop(CFRunLoopRef);
